pyproject.toml
[tool.poetry]
name = "DCNN_backend"
version = "0.1.0"
description = "backend for using ANN network"
authors = ["2662717402@qq.com"]
package-mode = false

[tool.poetry.dependencies]
python = "3.12.3" # 根据你的项目实际情况指定Python版本
Flask = "3.1.0"
pandas = "2.2.3"
joblib = "1.4.2"
requests = "2.32.3"
tensorflow = "2.18.0"
scikit-learn = "1.6.0"
keras = "3.8.0"
gunicorn = "^23.0.0"
mysql-connector-python = "^9.2.0"
